requirements.txt
fastapi
uvicorn
pandas
numpy
requests
duckdb
ta
scikit-learn
hmmlearn
torch
transformers
redis
python-dotenv
